move autocomplete code to own file
